Split CRUDRepository into separate Protocols
