+++ Example/User.swift
@@ -26,7 +26,7 @@ struct User: Identifiable {
     let email: String
 }
 
-extension User: DictionaryInitializable, DictionaryRepresentable {
+extension User: DictionaryInitializable {
     
     init(dictionary: [String: Any]) throws {
         guard let
@@ -38,13 +38,5 @@ extension User: DictionaryInitializable, DictionaryRepresentable {
         self.name = name
         self.email = email
     }
-    
-    var dictionaryValue: [String: Any] {
-        return [
-            "id": self.id,
-            "name": self.name,
-            "email": self.email
-        ]
-    }
-    
+
 }
